+++ src/main/java/racingcar/service/CarFactory.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.service;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.model.Car;
 
 public class CarFactory {
